+++ zk.js/src/idls/psp_compressed_pda_program.ts
@@ -3,7 +3,10 @@ export type PspCompressedPdaProgram = {
   "name": "psp_compressed_pda_program",
   "instructions": [
     {
-      "name": "initializeUserEntry",
+      "name": "executeCompressedTransaction",
+      "docs": [
+        "This function can be used to transfer sol and execute any other compressed transaction."
+      ],
       "accounts": [
         {
           "name": "signer",
@@ -11,46 +14,67 @@ export type PspCompressedPdaProgram = {
           "isSigner": true
         },
         {
-          "name": "systemProgram",
+          "name": "authorityPda",
+          "isMut": false,
+          "isSigner": false,
+          "docs": [
+            "Check that mint authority is derived from signer"
+          ]
+        },
+        {
+          "name": "registeredProgramPda",
+          "isMut": true,
+          "isSigner": false,
+          "docs": [
+            "CHECK this account"
+          ]
+        },
+        {
+          "name": "noopProgram",
+          "isMut": false,
+          "isSigner": false,
+          "docs": [
+            "CHECK this account"
+          ]
+        },
+        {
+          "name": "compressedPdaProgram",
           "isMut": false,
           "isSigner": false
         },
         {
-          "name": "userEntry",
+          "name": "pspAccountCompressionAuthority",
           "isMut": true,
-          "isSigner": false
+          "isSigner": false,
+          "docs": [
+            "CHECK this account in psp account compression program"
+          ]
+        },
+        {
+          "name": "accountCompressionProgram",
+          "isMut": false,
+          "isSigner": false,
+          "docs": [
+            "CHECK this account in psp account compression program"
+          ]
         }
       ],
       "args": [
         {
-          "name": "lightPubkey",
-          "type": {
-            "array": [
-              "u8",
-              32
-            ]
-          }
-        },
-        {
-          "name": "lightEncryptionPubkey",
-          "type": {
-            "array": [
-              "u8",
-              32
-            ]
-          }
+          "name": "inputs",
+          "type": "bytes"
         }
       ]
     }
   ],
   "accounts": [
     {
-      "name": "userEntry",
+      "name": "instructionDataTransfer",
       "type": {
         "kind": "struct",
         "fields": [
           {
-            "name": "solanaPubkey",
+            "name": "proofA",
             "type": {
               "array": [
                 "u8",
@@ -59,7 +83,16 @@ export type PspCompressedPdaProgram = {
             }
           },
           {
-            "name": "lightPubkey",
+            "name": "proofB",
+            "type": {
+              "array": [
+                "u8",
+                64
+              ]
+            }
+          },
+          {
+            "name": "proofC",
             "type": {
               "array": [
                 "u8",
@@ -68,12 +101,171 @@ export type PspCompressedPdaProgram = {
             }
           },
           {
-            "name": "lightEncryptionPubkey",
+            "name": "inUtxos",
+            "type": {
+              "vec": {
+                "defined": "TokenInUtxo"
+              }
+            }
+          },
+          {
+            "name": "lowElementIndexes",
+            "type": {
+              "vec": "u16"
+            }
+          },
+          {
+            "name": "rootIndexes",
+            "type": {
+              "vec": "u64"
+            }
+          },
+          {
+            "name": "rpcFee",
+            "type": {
+              "option": "u64"
+            }
+          },
+          {
+            "name": "outUtxo",
+            "type": {
+              "vec": {
+                "defined": "TokenOutUtxo"
+              }
+            }
+          }
+        ]
+      }
+    },
+    {
+      "name": "tokenInUtxo",
+      "type": {
+        "kind": "struct",
+        "fields": [
+          {
+            "name": "owner",
+            "type": "publicKey"
+          },
+          {
+            "name": "blinding",
             "type": {
               "array": [
                 "u8",
                 32
               ]
+            }
+          },
+          {
+            "name": "lamports",
+            "type": "u64"
+          },
+          {
+            "name": "data",
+            "type": {
+              "option": {
+                "defined": "TlvData"
+              }
+            }
+          }
+        ]
+      }
+    },
+    {
+      "name": "tokenOutUtxo",
+      "type": {
+        "kind": "struct",
+        "fields": [
+          {
+            "name": "owner",
+            "type": "publicKey"
+          },
+          {
+            "name": "blinding",
+            "type": {
+              "array": [
+                "u8",
+                32
+              ]
+            }
+          },
+          {
+            "name": "lamports",
+            "type": "u64"
+          },
+          {
+            "name": "data",
+            "type": {
+              "option": {
+                "defined": "TlvData"
+              }
+            }
+          }
+        ]
+      }
+    }
+  ],
+  "types": [
+    {
+      "name": "TlvData",
+      "docs": [
+        "Time lock escrow example:",
+        "escrow tlv data -> compressed token program",
+        "let escrow_data = {",
+        "owner: Pubkey, // owner is the user pubkey",
+        "release_slot: u64,",
+        "deposit_slot: u64,",
+        "};",
+        "",
+        "let escrow_tlv_data = TlvData {",
+        "discriminator: [1,0,0,0,0,0,0,0],",
+        "owner: escrow_program_id,",
+        "data: escrow_data,",
+        "tlv_data: Some(token_tlv.try_to_vec()?),",
+        "};",
+        "let token_tlv = TlvData {",
+        "discriminator: [2,0,0,0,0,0,0,0],",
+        "owner: token_program,",
+        "data: token_data,",
+        "tlv_data: None,",
+        "};",
+        "let token_data = TokenAccount {",
+        "mint,",
+        "owner,",
+        "amount: 10_000_000u64,",
+        "delegate: None,",
+        "state: Initialized, (u64)",
+        "is_native: None,",
+        "delegated_amount: 0u64,",
+        "close_authority: None,",
+        "};",
+        ""
+      ],
+      "type": {
+        "kind": "struct",
+        "fields": [
+          {
+            "name": "discriminator",
+            "type": {
+              "array": [
+                "u8",
+                8
+              ]
+            }
+          },
+          {
+            "name": "owner",
+            "type": "publicKey"
+          },
+          {
+            "name": "data",
+            "type": "bytes"
+          },
+          {
+            "name": "tlvData",
+            "type": {
+              "option": {
+                "defined": "Box"
+              }
             }
           }
         ]
@@ -87,7 +279,10 @@ export const IDL: PspCompressedPdaProgram = {
   "name": "psp_compressed_pda_program",
   "instructions": [
     {
-      "name": "initializeUserEntry",
+      "name": "executeCompressedTransaction",
+      "docs": [
+        "This function can be used to transfer sol and execute any other compressed transaction."
+      ],
       "accounts": [
         {
           "name": "signer",
@@ -95,46 +290,67 @@ export const IDL: PspCompressedPdaProgram = {
           "isSigner": true
         },
         {
-          "name": "systemProgram",
+          "name": "authorityPda",
+          "isMut": false,
+          "isSigner": false,
+          "docs": [
+            "Check that mint authority is derived from signer"
+          ]
+        },
+        {
+          "name": "registeredProgramPda",
+          "isMut": true,
+          "isSigner": false,
+          "docs": [
+            "CHECK this account"
+          ]
+        },
+        {
+          "name": "noopProgram",
+          "isMut": false,
+          "isSigner": false,
+          "docs": [
+            "CHECK this account"
+          ]
+        },
+        {
+          "name": "compressedPdaProgram",
           "isMut": false,
           "isSigner": false
         },
         {
-          "name": "userEntry",
+          "name": "pspAccountCompressionAuthority",
           "isMut": true,
-          "isSigner": false
+          "isSigner": false,
+          "docs": [
+            "CHECK this account in psp account compression program"
+          ]
+        },
+        {
+          "name": "accountCompressionProgram",
+          "isMut": false,
+          "isSigner": false,
+          "docs": [
+            "CHECK this account in psp account compression program"
+          ]
         }
       ],
       "args": [
         {
-          "name": "lightPubkey",
-          "type": {
-            "array": [
-              "u8",
-              32
-            ]
-          }
-        },
-        {
-          "name": "lightEncryptionPubkey",
-          "type": {
-            "array": [
-              "u8",
-              32
-            ]
-          }
+          "name": "inputs",
+          "type": "bytes"
         }
       ]
     }
   ],
   "accounts": [
     {
-      "name": "userEntry",
+      "name": "instructionDataTransfer",
       "type": {
         "kind": "struct",
         "fields": [
           {
-            "name": "solanaPubkey",
+            "name": "proofA",
             "type": {
               "array": [
                 "u8",
@@ -143,7 +359,16 @@ export const IDL: PspCompressedPdaProgram = {
             }
           },
           {
-            "name": "lightPubkey",
+            "name": "proofB",
+            "type": {
+              "array": [
+                "u8",
+                64
+              ]
+            }
+          },
+          {
+            "name": "proofC",
             "type": {
               "array": [
                 "u8",
@@ -152,12 +377,171 @@ export const IDL: PspCompressedPdaProgram = {
             }
           },
           {
-            "name": "lightEncryptionPubkey",
+            "name": "inUtxos",
+            "type": {
+              "vec": {
+                "defined": "TokenInUtxo"
+              }
+            }
+          },
+          {
+            "name": "lowElementIndexes",
+            "type": {
+              "vec": "u16"
+            }
+          },
+          {
+            "name": "rootIndexes",
+            "type": {
+              "vec": "u64"
+            }
+          },
+          {
+            "name": "rpcFee",
+            "type": {
+              "option": "u64"
+            }
+          },
+          {
+            "name": "outUtxo",
+            "type": {
+              "vec": {
+                "defined": "TokenOutUtxo"
+              }
+            }
+          }
+        ]
+      }
+    },
+    {
+      "name": "tokenInUtxo",
+      "type": {
+        "kind": "struct",
+        "fields": [
+          {
+            "name": "owner",
+            "type": "publicKey"
+          },
+          {
+            "name": "blinding",
             "type": {
               "array": [
                 "u8",
                 32
               ]
+            }
+          },
+          {
+            "name": "lamports",
+            "type": "u64"
+          },
+          {
+            "name": "data",
+            "type": {
+              "option": {
+                "defined": "TlvData"
+              }
+            }
+          }
+        ]
+      }
+    },
+    {
+      "name": "tokenOutUtxo",
+      "type": {
+        "kind": "struct",
+        "fields": [
+          {
+            "name": "owner",
+            "type": "publicKey"
+          },
+          {
+            "name": "blinding",
+            "type": {
+              "array": [
+                "u8",
+                32
+              ]
+            }
+          },
+          {
+            "name": "lamports",
+            "type": "u64"
+          },
+          {
+            "name": "data",
+            "type": {
+              "option": {
+                "defined": "TlvData"
+              }
+            }
+          }
+        ]
+      }
+    }
+  ],
+  "types": [
+    {
+      "name": "TlvData",
+      "docs": [
+        "Time lock escrow example:",
+        "escrow tlv data -> compressed token program",
+        "let escrow_data = {",
+        "owner: Pubkey, // owner is the user pubkey",
+        "release_slot: u64,",
+        "deposit_slot: u64,",
+        "};",
+        "",
+        "let escrow_tlv_data = TlvData {",
+        "discriminator: [1,0,0,0,0,0,0,0],",
+        "owner: escrow_program_id,",
+        "data: escrow_data,",
+        "tlv_data: Some(token_tlv.try_to_vec()?),",
+        "};",
+        "let token_tlv = TlvData {",
+        "discriminator: [2,0,0,0,0,0,0,0],",
+        "owner: token_program,",
+        "data: token_data,",
+        "tlv_data: None,",
+        "};",
+        "let token_data = TokenAccount {",
+        "mint,",
+        "owner,",
+        "amount: 10_000_000u64,",
+        "delegate: None,",
+        "state: Initialized, (u64)",
+        "is_native: None,",
+        "delegated_amount: 0u64,",
+        "close_authority: None,",
+        "};",
+        ""
+      ],
+      "type": {
+        "kind": "struct",
+        "fields": [
+          {
+            "name": "discriminator",
+            "type": {
+              "array": [
+                "u8",
+                8
+              ]
+            }
+          },
+          {
+            "name": "owner",
+            "type": "publicKey"
+          },
+          {
+            "name": "data",
+            "type": "bytes"
+          },
+          {
+            "name": "tlvData",
+            "type": {
+              "option": {
+                "defined": "Box"
+              }
             }
           }
         ]
