+++ zk.js/src/idls/psp_compressed_token_program.ts
@@ -340,140 +340,6 @@ export type PspCompressedTokenProgram = {
           }
         ]
       }
-    },
-    {
-      "name": "tokenInUtxo",
-      "type": {
-        "kind": "struct",
-        "fields": [
-          {
-            "name": "owner",
-            "type": "publicKey"
-          },
-          {
-            "name": "blinding",
-            "type": {
-              "array": [
-                "u8",
-                32
-              ]
-            }
-          },
-          {
-            "name": "lamports",
-            "type": "u64"
-          },
-          {
-            "name": "data",
-            "type": {
-              "option": {
-                "defined": "TlvData"
-              }
-            }
-          }
-        ]
-      }
-    },
-    {
-      "name": "tokenOutUtxo",
-      "type": {
-        "kind": "struct",
-        "fields": [
-          {
-            "name": "owner",
-            "type": "publicKey"
-          },
-          {
-            "name": "blinding",
-            "type": {
-              "array": [
-                "u8",
-                32
-              ]
-            }
-          },
-          {
-            "name": "lamports",
-            "type": "u64"
-          },
-          {
-            "name": "data",
-            "type": {
-              "option": {
-                "defined": "TlvData"
-              }
-            }
-          }
-        ]
-      }
-    }
-  ],
-  "types": [
-    {
-      "name": "TlvData",
-      "docs": [
-        "Time lock escrow example:",
-        "escrow tlv data -> compressed token program",
-        "let escrow_data = {",
-        "owner: Pubkey, // owner is the user pubkey",
-        "release_slot: u64,",
-        "deposit_slot: u64,",
-        "};",
-        "",
-        "let escrow_tlv_data = TlvData {",
-        "discriminator: [1,0,0,0,0,0,0,0],",
-        "owner: escrow_program_id,",
-        "data: escrow_data,",
-        "tlv_data: Some(token_tlv.try_to_vec()?),",
-        "};",
-        "let token_tlv = TlvData {",
-        "discriminator: [2,0,0,0,0,0,0,0],",
-        "owner: token_program,",
-        "data: token_data,",
-        "tlv_data: None,",
-        "};",
-        "let token_data = TokenAccount {",
-        "mint,",
-        "owner,",
-        "amount: 10_000_000u64,",
-        "delegate: None,",
-        "state: Initialized, (u64)",
-        "is_native: None,",
-        "delegated_amount: 0u64,",
-        "close_authority: None,",
-        "};",
-        ""
-      ],
-      "type": {
-        "kind": "struct",
-        "fields": [
-          {
-            "name": "discriminator",
-            "type": {
-              "array": [
-                "u8",
-                8
-              ]
-            }
-          },
-          {
-            "name": "owner",
-            "type": "publicKey"
-          },
-          {
-            "name": "data",
-            "type": "bytes"
-          },
-          {
-            "name": "tlvData",
-            "type": {
-              "option": {
-                "defined": "Box"
-              }
-            }
-          }
-        ]
-      }
     }
   ],
   "errors": [
@@ -827,140 +693,6 @@ export const IDL: PspCompressedTokenProgram = {
             "type": {
               "vec": {
                 "defined": "TokenOutUtxo"
-              }
-            }
-          }
-        ]
-      }
-    },
-    {
-      "name": "tokenInUtxo",
-      "type": {
-        "kind": "struct",
-        "fields": [
-          {
-            "name": "owner",
-            "type": "publicKey"
-          },
-          {
-            "name": "blinding",
-            "type": {
-              "array": [
-                "u8",
-                32
-              ]
-            }
-          },
-          {
-            "name": "lamports",
-            "type": "u64"
-          },
-          {
-            "name": "data",
-            "type": {
-              "option": {
-                "defined": "TlvData"
-              }
-            }
-          }
-        ]
-      }
-    },
-    {
-      "name": "tokenOutUtxo",
-      "type": {
-        "kind": "struct",
-        "fields": [
-          {
-            "name": "owner",
-            "type": "publicKey"
-          },
-          {
-            "name": "blinding",
-            "type": {
-              "array": [
-                "u8",
-                32
-              ]
-            }
-          },
-          {
-            "name": "lamports",
-            "type": "u64"
-          },
-          {
-            "name": "data",
-            "type": {
-              "option": {
-                "defined": "TlvData"
-              }
-            }
-          }
-        ]
-      }
-    }
-  ],
-  "types": [
-    {
-      "name": "TlvData",
-      "docs": [
-        "Time lock escrow example:",
-        "escrow tlv data -> compressed token program",
-        "let escrow_data = {",
-        "owner: Pubkey, // owner is the user pubkey",
-        "release_slot: u64,",
-        "deposit_slot: u64,",
-        "};",
-        "",
-        "let escrow_tlv_data = TlvData {",
-        "discriminator: [1,0,0,0,0,0,0,0],",
-        "owner: escrow_program_id,",
-        "data: escrow_data,",
-        "tlv_data: Some(token_tlv.try_to_vec()?),",
-        "};",
-        "let token_tlv = TlvData {",
-        "discriminator: [2,0,0,0,0,0,0,0],",
-        "owner: token_program,",
-        "data: token_data,",
-        "tlv_data: None,",
-        "};",
-        "let token_data = TokenAccount {",
-        "mint,",
-        "owner,",
-        "amount: 10_000_000u64,",
-        "delegate: None,",
-        "state: Initialized, (u64)",
-        "is_native: None,",
-        "delegated_amount: 0u64,",
-        "close_authority: None,",
-        "};",
-        ""
-      ],
-      "type": {
-        "kind": "struct",
-        "fields": [
-          {
-            "name": "discriminator",
-            "type": {
-              "array": [
-                "u8",
-                8
-              ]
-            }
-          },
-          {
-            "name": "owner",
-            "type": "publicKey"
-          },
-          {
-            "name": "data",
-            "type": "bytes"
-          },
-          {
-            "name": "tlvData",
-            "type": {
-              "option": {
-                "defined": "Box"
               }
             }
           }
